处理队标

+++ handle/management/commands/handle_avatar.py
@@ -23,7 +23,7 @@ class Command(BaseCommand):
             print(host_avatar, '===', host_name + '.' + host_avatar_type)
             print(guest_avatar, '===', guest_name + '.' + guest_avatar_type)
 
-            if os.path.exists(host_name + '.' + host_avatar_type) is not True:
+            if os.path.exists(host_name + '.' + host_avatar_type) is not True and os.path.exists(host_avatar) is True:
                 shutil.copy(host_avatar, host_name + '.' + host_avatar_type)
-            if os.path.exists(guest_name + '.' + guest_avatar_type) is not True:
+            if os.path.exists(guest_name + '.' + guest_avatar_type) is not True and os.path.exists(guest_avatar) is True:
                 shutil.copy(guest_avatar, guest_name + '.' + guest_avatar_type)
